arm64: read enable-method for CPU0
With the advent of CPU_HOTPLUG, the enable-method property for CPU0 may
tells us something useful (i.e. how to hotplug it back on), so we must
read it along with all the enable-method for all the other CPUs.  Even
on UP the enable-method may tell us useful information (e.g. if a core
has some mechanism that might be usable for cpuidle), so we should
always read it.

This patch factors out the reading of the enable method, and ensures
that CPU0's enable method is read regardless of whether the kernel is
built with SMP support.

Bug 2064472

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Change-Id: Ibf6b27bf88e0cc5262a8b08d2f2ba2aa09953c40
Reviewed-on: http://git-master/r/365942
Signed-off-by: Bo Yan <byan@nvidia.com>
Signed-off-by: Prashant Gaikwad <pgaikwad@nvidia.com>
(cherry picked from commit a33318ea71dc1f3706c4b315dfe6b4a90ed2f82a in rel-24-uda-r1)
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1740691
Reviewed-by: Shreshtha Sahu <ssahu@nvidia.com>
GVS: Gerrit_Virtual_Submit