arm64: alternative: Provide if/else/endif assembler macros
The existing alternative_insn macro has some limitations that make it
hard to work with. In particular the fact it takes instructions from it
own macro arguments means it doesn't play very nicely with C pre-processor
macros because the macro arguments look like a string to the C
pre-processor. Workarounds are (probably) possible but things start to
look ugly.

Introduce an alternative set of macros that allows instructions to be
presented to the assembler as normal and switch everything over to the
new macros.

==  This is a back port change from K4.4 to K3.10 so it also includes: ==
  arm64: alternatives: add enable parameter to conditional asm macros

  There are cases where we want to compile out both versions of an
  alternative code block, so add an enable parameter to the new conditional
  alternative assembly macros in the same way as alternative_insn.

Bug 1975157

Change-Id: I39fd42525f717c63e9b5f8a9ec182e77a3e28401
Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Gagan Grover <ggrover@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1626807
(cherry picked from commit 4ae908187dee327273f159a34ae11b8516421d57)
Signed-off-by: Jeetesh Burman <jburman@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1648617
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
GVS: Gerrit_Virtual_Submit