[net]: Move facnout check inside lock
In __unregister_prot_hook and packet_release of af_packet.c, there is a possible use-after-free due to improper locking. This could lead to local escalation of privilege in the kernel with System execution privileges needed. User interaction is not needed for exploitation
for CVE-2018-9439
Bug 2240432

Change-Id: I5a4325df2c1162746c1153f948bb93577547f4a0
Signed-off-by: Amulya Yarlagadda <ayarlagadda@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1788934
Reviewed-on: https://git-master.nvidia.com/r/1802102
GVS: Gerrit_Virtual_Submit
Reviewed-by: Prabhu Kuttiyam <pkuttiyam@nvidia.com>